refactor(make apic generic) make apic unaware of bauhaus and split in 3 projects
