+++ apic/src/main/resources/application.properties
@@ -1,0 +1,16 @@
+spring.config.import=optional:file:${HOME}/props/local-apic.properties
+fr.insee.apic.backend.url=http://myapp.insee.fr
+fr.insee.apic.cors.allowed-origins=http://myapp-client.insee.fr
+spring.threads.virtual.enabled=true
+server.tomcat.accesslog.enabled=true
+server.tomcat.accesslog.directory=${HOME}/logs
+server.http2.enabled=true
+
+fr.insee.apic.interface-controllers.package=fr.insee.myapp.parent_package_of_controllers
+
+jwt.role-claim = realm_access
+jwt.role-claim.roles = roles
+jwt.id-claim= preferred_username
+spring.security.oauth2.resourceserver.jwt.issuer-uri=${fr.insee.apic.auth-server-url}
+
+fr.insee.apic.security.csrf.disable=false
