+++ bauhaus/src/main/java/fr/insee/rmes/bauhaus/model/OperationsLink.java
@@ -1,0 +1,4 @@
+package fr.insee.rmes.bauhaus.model;
+
+public record OperationsLink(String id, String type, String labelLg1, String LabelLg2) {
+}
