+++ sample-and-tests/src/test/java/fr/insee/rmes/integration/authorizations/TokenForTestsConfiguration.java
