--- src/main/java/fr/insee/rmes/ApiConsultationPourBauhausBackOfficeApplication.java
@@ -1,20 +1,0 @@
-package fr.insee.rmes;
-
-import fr.insee.rmes.config.ControllersConfiguration;
-import fr.insee.rmes.config.PropertiesLogger;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-
-@SpringBootApplication
-public class ApiConsultationPourBauhausBackOfficeApplication {
-
-	public static void main(String[] args) {
-		configureApplicationBuilder(new SpringApplicationBuilder()).build().run(args);
-	}
-
-	private static SpringApplicationBuilder configureApplicationBuilder(SpringApplicationBuilder springApplicationBuilder) {
-		return springApplicationBuilder.sources(ApiConsultationPourBauhausBackOfficeApplication.class)
-				.listeners(new PropertiesLogger(), new ControllersConfiguration());
-	}
-
-}
