--- src/main/java/fr/insee/rmes/model/OperationsLink.java
@@ -1,4 +1,0 @@
-package fr.insee.rmes.model;
-
-public record OperationsLink(String id, String type, String labelLg1, String LabelLg2) {
-}
