--- src/main/resources/application.properties
@@ -1,16 +1,0 @@
-spring.config.import=optional:file:${HOME}/props/rmes-apic.properties
-fr.insee.rmes.apic.bauhaus-bo.url=http://bauhaus.insee.fr
-#fr.insee.rmes.apic.cors.allowed-origins=http://bauhaus-front.insee.fr
-spring.threads.virtual.enabled=true
-server.tomcat.accesslog.enabled=true
-server.tomcat.accesslog.directory=${HOME}/logs
-server.http2.enabled=true
-
-fr.insee.rmes.interface-controllers.package=fr.insee.rmes.bauhauscontrollers
-
-jwt.role-claim = realm_access
-jwt.role-claim.roles = roles
-jwt.id-claim= preferred_username
-spring.security.oauth2.resourceserver.jwt.issuer-uri=${fr.insee.rmes.bauhaus.auth-server-url}
-
-fr.insee.rmes.security.csrf.disable=false
