--- src/main/resources/schema.sql
@@ -1,1 +1,0 @@
-create table if not exists my_pojo (id bigint GENERATED ALWAYS AS IDENTITY, value_string varchar(100) not null);
