--- src/test/java/fr/insee/rmes/ApiConsultationPourBauhausBackOfficeApplicationTests.java
@@ -1,13 +1,0 @@
-package fr.insee.rmes;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiConsultationPourBauhausBackOfficeApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
